Move stack widget to pyqt/extended
